Adding support for module configuration
This adds support for creating an isolated, templated, set of configuration options specific to this module.

This is then loaded into the default apache config file (using line in file). This is a loose dependency as this role does not to keep track of the main config file and the apache role does not need to know this role will be later be used.